Consolidate sign_in spec support files
