+++ spec/rails_helper.rb
@@ -30,8 +30,8 @@ require 'support/pdf_fill_helper'
 require 'support/vcr_multipart_matcher_helper'
 require 'support/request_helper'
 require 'support/uploader_helpers'
-require 'support/service_account_authorization_context'
 require 'support/sign_in/certifiable_shared_examples'
+require 'support/sign_in/service_account_authorization_context'
 require 'super_diff/rspec-rails'
 require 'super_diff/active_support'
 require './spec/support/default_configuration_helper'
